# Optimize sql_fingerprint() a bit